+++ Kernel/System/AuthSession.pm
@@ -19,6 +19,7 @@ our @ObjectDependencies = (
     'Kernel::System::Log',
     'Kernel::System::Main',
     'Kernel::System::SystemData',
+    'Kernel::System::SysConfig',
 );
 
 =head1 NAME
@@ -241,6 +242,8 @@ sub CreateSessionID {
     my $SessionLimit;
     if ( $Param{UserType} eq 'User' ) {
 
+        # Use the AgentSessionLimit from the business solution, if a session limit exists and use the AgentSessionLimit
+        #   from the config, if the value is lower the business solution value.
         $SessionLimit = $OTRSBusinessSystemData{AgentSessionLimit};
         if ( !$SessionLimit || ( $Self->{AgentSessionLimit} && $Self->{AgentSessionLimit} < $SessionLimit ) ) {
             $SessionLimit = $Self->{AgentSessionLimit};
@@ -249,6 +252,7 @@ sub CreateSessionID {
     elsif ( $Param{UserType} eq 'Customer' && $Self->{CustomerSessionLimit} ) {
         $SessionLimit = $Self->{CustomerSessionLimit};
     }
+
 
     # get session per user limit config
     my $SessionPerUserLimit;
@@ -259,7 +263,10 @@ sub CreateSessionID {
         $SessionPerUserLimit = $Self->{CustomerSessionPerUserLimit};
     }
 
-    if ( $SessionLimit || $SessionPerUserLimit ) {
+    my $SessionSource = $Param{SessionSource} || '';
+
+    # Don't check the session limit for sessions from the source 'GenericInterface'.
+    if ( $SessionSource ne 'GenericInterface' && ( $SessionLimit || $SessionPerUserLimit ) ) {
 
         my %ActiveSessions = $Self->GetActiveSessions(%Param);