+++ Kernel/System/AuthSession/DB.pm
@@ -393,6 +393,7 @@ sub GetActiveSessions {
             WHERE data_key = 'UserType'
                 OR data_key = 'UserLastRequest'
                 OR data_key = 'UserLogin'
+                OR data_key = 'SessionSource'
             ORDER BY id ASC",
     );
 
@@ -413,6 +414,11 @@ sub GetActiveSessions {
 
         next SESSIONID if !$SessionID;
         next SESSIONID if !$SessionData{$SessionID};
+
+        # Don't count session from source 'GenericInterface'
+        my $SessionSource = $SessionData{$SessionID}->{SessionSource} || '';
+
+        next SESSIONID if $SessionSource eq 'GenericInterface';
 
         # get needed data
         my $UserType        = $SessionData{$SessionID}->{UserType}        || '';