+++ Kernel/System/AuthSession/FS.pm
@@ -248,8 +248,13 @@ sub CreateSessionID {
     my $UserLogin        = $Self->{Cache}->{$SessionID}->{UserLogin}        || '';
     my $UserSessionStart = $Self->{Cache}->{$SessionID}->{UserSessionStart} || '';
     my $UserLastRequest  = $Self->{Cache}->{$SessionID}->{UserLastRequest}  || '';
+    my $SessionSource    = $Self->{Cache}->{$SessionID}->{SessionSource}    || '';
 
     my $StateContent = $UserType . '####' . $UserLogin . '####' . $UserSessionStart . '####' . $UserLastRequest;
+
+    if ($SessionSource) {
+        $StateContent .= '####' . $SessionSource;
+    }
 
     # write state file
     $MainObject->FileWrite(
@@ -394,6 +399,10 @@ sub GetActiveSessions {
         my $UserType        = $SessionData[0] || '';
         my $UserLogin       = $SessionData[1] || '';
         my $UserLastRequest = $SessionData[3] || $TimeNow;
+        my $SessionSource   = $SessionData[4] || '';
+
+        # Don't count sessions from source 'GenericInterface'.
+        next SESSIONID if $SessionSource eq 'GenericInterface';
 
         next SESSIONID if $UserType ne $Param{UserType};
 
@@ -541,11 +550,16 @@ sub DESTROY {
         my $UserLogin        = $Self->{Cache}->{$SessionID}->{UserLogin}        || '';
         my $UserSessionStart = $Self->{Cache}->{$SessionID}->{UserSessionStart} || '';
         my $UserLastRequest  = $Self->{Cache}->{$SessionID}->{UserLastRequest}  || '';
+        my $SessionSource    = $Self->{Cache}->{$SessionID}->{SessionSource}    || '';
 
         my $StateContent = $UserType . '####'
             . $UserLogin . '####'
             . $UserSessionStart . '####'
             . $UserLastRequest;
+
+        if ($SessionSource) {
+            $StateContent .= '####' . $SessionSource;
+        }
 
         # write state file
         $MainObject->FileWrite(